Refactored code to include dispatcher classes.
