+++ openequivariance/implementations/convolution/CUEConv.py
@@ -2,10 +2,10 @@ import numpy as np
 import numpy.linalg as la
 
 from openequivariance.implementations.CUETensorProduct import CUETensorProduct
-from openequivariance.implementations.convolution.Convolution import *
+from openequivariance.implementations.convolution.ConvolutionBase import *
 from openequivariance.benchmark.tpp_creation_utils import *
 
-class CUEConv(Convolution):
+class CUEConv(ConvolutionBase):
     def __init__(self, config, idx_dtype=np.int64, torch_op=True):
         super().__init__(config, idx_dtype, torch_op)
