+++ openequivariance/implementations/convolution/NumpyConv.py
@@ -1,10 +1,10 @@
 import numpy as np
 import numpy.linalg as la
 
-from openequivariance.implementations.convolution.Convolution import *
+from openequivariance.implementations.convolution.ConvolutionBase import *
 from openequivariance.implementations.NumpyTensorProduct import *
 
-class NumpyConv(Convolution):
+class NumpyConv(ConvolutionBase):
     def __init__(self, config):
         self.config = config 
         self.L1, self.L2, self.L3 = config.irreps_in1, config.irreps_in2, config.irreps_out
