+++ openequivariance/implementations/convolution/TensorProductConv.py
@@ -1,0 +1,10 @@
+from openequivariance.implementations.convolution.LoopUnrollConv import *
+
+class TensorProductConv(LoopUnrollConv):
+    '''
+    Dispatcher class for convolutions. Right now, we just subclass
+    LoopUnrollConv.
+    '''
+    def __init__(self, config, torch_op=False, deterministic=False):
+        super().__init__(config, idx_dtype=np.int64,
+                torch_op=torch_op, deterministic=deterministic)
